move getUpvotesStatus and getUserData to helperFunctions file
